+++ test/aeso_parser_tests.erl
@@ -17,7 +17,7 @@ simple_contracts_test_() ->
             ?assertMatch(
                 [{contract_main, _, {con, _, "Identity"},
                     [{letfun, _, {id, _, "id"}, [{id, _, "x"}], {id, _, "_"},
-                        {id, _, "x"}}]}], parse_string(Text)),
+                        [{guarded, _, [], {id, _, "x"}}]}]}], parse_string(Text)),
             ok
        end},
       {"Operator precedence test.",